Added ResourceSystem
In order to facilitate testing the very simple non-functional TextFile resource was created.
This resource does nothing, but will allow other's to create resources based on it. Also it is
needed for the tests.

Reflection templates have been added to Trillek.h. These are mostly useful for resource
types, but other classes can use them for example for streaming of the network.

The tests CMakeLists was modified to adjust the binary output path and to add the
src/ directory under tests for those tests that require actual source files. This was done to
selectively and manually added specific source files, and not just add every single one
in the project.

Also fixed the includes in PropertyTest.h and changed include guards from #pragma to #ifdef.