Added Material and Shader resources (missed in the last commit)
These are useful for rendering only.

Material stores a list of pairs for texture to texture IDs that can be looped over during
rendering and for pre-sorted Renderables for rendering.

Shader loads GLSL files and compiles them and has methods for using/unusing,
assigning uniforms/attributes, and retrieving the uniform/attribute locations.