Added RegisterTypes method for ResrouceSystem and JSONParser
These methods couple the registered types at compile time and are
called from the appropriate construction calls. The only onus for the
developer is to add the appropriate include and registration line to
the correct source files and their type is then known to the corresponding
system.

Additionally some GL libraries have been added to the tests CMakeLists
as certain files require them.