Added event dispatcher
This is a template based class that is subscribed to via
GetInstance()->Subscribe(entity_id, Subscriber<T>*)
where type is the type you want to be notified of changes
to. The type T needs to call GetInstance()->NotifySubscribers(entity_id, T* this)
to notify the subscribers of a change. Subscribers just
implement the Subscriber interface.

Also updated the submodule.