Renderables no longer know about materials
Materials are stored in material groups only.
Renderables are compared for grouping in the
material group by their shader.

Renderables store a list of Texture(s) in each
buffer group. On UpdateBufferGroups the
textures are added loaded and added to the
buffer groups.

RenderSystem::AddComponent takes the texture
list and adds the textures to the material group's
material.