Layer management functions and various graphics system changes
The clear function on RenderLayer is changed to switch draw layers less.
Texture params for multisample textures are not needed and removed.
command resolver uses the new Container class and RTTI features to
determine types in Property and Container items. Layer bind and copy is
implemented, allowing control of offscreen rendering (important)