Renderable now uses dynamic textures by default
Fixed Texture::Load() for dynamic textures to use the source format
instead of just using GL:RGBA (not every texture uses this format.)

The property dynamic_texture can now be specified for renderable
components (true by default). What this does is causes the PixelBuffer
for each texture to get a unique name (entityID_filename) that is unique
to this renderable instance and will be updated each frame. To have a
renderable share its textures with other instance just set dynamic_texture
to false in the JSON config.